credit_approval/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
celer==0.7.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
download==0.3.5
drf-yasg==1.21.10
et_xmlfile==2.0.0
fonttools==4.59.0
idna==3.10
inflection==0.5.1
joblib==1.5.1
kiwisolver==1.4.8
libsvmdata==0.4.1
matplotlib==3.10.5
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xarray==2025.6.1
gunicorn